2차 코드 리뷰

+++ point/src/main/java/store/_0982/point/application/dto/MemberPointInfo.java
@@ -1,13 +1,13 @@
 package store._0982.point.application.dto;
 
-import store._0982.point.domain.MemberPoint;
+import store._0982.point.domain.entity.MemberPoint;
 
 import java.time.OffsetDateTime;
 import java.util.UUID;
 
 public record MemberPointInfo(
         UUID memberId,
-        int pointBalance,
+        long pointBalance,
         OffsetDateTime lastUsedAt
 ) {
     public static MemberPointInfo from(MemberPoint memberPoint){
