+++ point/src/main/java/store/_0982/point/client/dto/TossPaymentCancelRequest.java
@@ -1,11 +1,11 @@
 package store._0982.point.client.dto;
 
 import store._0982.point.application.dto.PointRefundCommand;
-import store._0982.point.domain.PaymentPoint;
+import store._0982.point.domain.entity.PaymentPoint;
 
 public record TossPaymentCancelRequest(
         String paymentKey,
-        int amount,
+        long amount,
         String reason
 ) {
     public static TossPaymentCancelRequest from(PaymentPoint paymentPoint, PointRefundCommand command) {
